arreglo repositorios

+++ src/main/java/co/edu/udistrital/mdp/pets/repositories/ReportRepository.java
@@ -4,11 +4,10 @@ import co.edu.udistrital.mdp.pets.entities.ReportEntity;
 import co.edu.udistrital.mdp.pets.entities.ShelterEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 
 @Repository
-public interface ReportRepository extends JpaRepository<ReportEntity, Integer> {
+public interface ReportRepository extends JpaRepository<ReportEntity, Long> {
 
     List<ReportEntity> findByShelter(ShelterEntity shelter);
